Update CompiledCode.m
The only thing that this branch has extra is that it now has two for loops - the exact same for loop of the original, but twice - essentially, the idea is that the first for loop has a bigger stepSize, so after sweeping through the range of e and mu it would return the approximate optimal e and mu. The second for loop takes as an input these e and mu, and has a finer stepSize set around these e and mu. It does the same thing, but overall should give a slightly more precise e and mu. In theory, these two for loops should run faster than the single for loop since we can avoid sweeping through unimportant areas!