Impact angle vs width - No cap
This is a script that optimizes for Pn, Pt, and torque, then divides by torque to find the optimal width. I put no cap on the range of widths. At the end this graphs the widths vs impact angle. This runs alongside finderror.m, constraint.m and ellipse uniform data that are also included in this folder. You run the file angleVsWidth.m